Made subfolders for every section.
